% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dimensions_token.R
\name{get_dimensions_token}
\alias{get_dimensions_token}
\title{Get Dimensions db authorisation token}
\usage{
get_dimensions_token(username = NULL, password = NULL)
}
\arguments{
\item{username}{string}

\item{password}{string}
}
\value{
string
}
\description{
Get Dimensions db authorisation token
}
\examples{
\dontrun{
get_dimensions_token()
}

}
